Cleaner implementation
Added AbstractWebSocket for a cleaner implementation
